split `traits::prefix` into sub-modules
